fix : 실수 부동수소점 오차 이슈 해결

+++ src/main/java/calculator/view/output/MenuHistoryOutput.java
@@ -2,6 +2,7 @@ package calculator.view.output;
 
 import calculator.calculator.operator.Operators;
 
+import java.math.BigDecimal;
 import java.util.Collection;
 import java.util.HashMap;
 import java.util.Map;
@@ -12,10 +13,10 @@ import static java.util.stream.IntStream.range;
 public class MenuHistoryOutput implements BaseOutput {
 
     public void printHistories(Collection<?> histories) {
-        HashMap<String, Double> newHistories = new HashMap<>();
+        HashMap<String, BigDecimal> newHistories = new HashMap<>();
 
         histories.stream()
-                .map(history -> (Map.Entry<String, Double>) history)
+                .map(history -> (Map.Entry<String, BigDecimal>) history)
                 .forEach(parsedHistory -> newHistories.put(
                         parsedHistory.getKey(),
                         parsedHistory.getValue()));
@@ -23,7 +24,7 @@ public class MenuHistoryOutput implements BaseOutput {
         print(createHistories(newHistories) + ENTER.unit);
     }
 
-    private String createHistories(HashMap<String, Double> histories) {
+    private String createHistories(HashMap<String, BigDecimal> histories) {
         StringBuilder textBuilder = new StringBuilder();
 
         histories.keySet()
@@ -38,7 +39,7 @@ public class MenuHistoryOutput implements BaseOutput {
         return textBuilder.toString();
     }
 
-    private static void handleAnswer(HashMap<String, Double> histories, StringBuilder textBuilder, String formula, int idx) {
+    private static void handleAnswer(HashMap<String, BigDecimal> histories, StringBuilder textBuilder, String formula, int idx) {
         if (idx == formula.length() - 1) {
             textBuilder.append(SPACE.unit)
                     .append(EQUAL.unit)
